Compat-free!
[ Firmware ]
Removed dependency on pico-arduino-compat ( bleah )
Removed some test files
Did ( some ) cleanup of the main program
Rewrote the different loops
Added initial ping to genie, in order to validate it is operational
Decreased sleep times, as the ported genie library behaves and performs much better than the compat layer one
Added more debug prints

[ Display ]
Started redesign of the UI, to achieve a cleaner version
Note: this new version is not ( yet ) compatible with the firmware, as object ids have changed. FIXME is in the works.